change i32 to i64

+++ fibonacci/c.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
+#include <stdint.h>
 
-int fib(int n) {
+int64_t fib(int64_t n) {
     if (n <= 1) return n;
     return fib(n - 1) + fib(n - 2);
 }
